refactor: move relayout logic into dm-core
Moves Candidate/plan()/apply_candidates() from crates/cli into a new
dm_core::relayout module, so the TUI (and any future frontend) can compute
and apply layout drift without depending on the cli crate. Display/prompt
code (print_candidates, the confirm) stays in the cli crate.

apply_candidates now returns skipped target names instead of eprintln-ing
them directly, since core shouldn't have console side effects — callers
report skips however fits (plain println in the CLI, a status line later
in the TUI).

Pure refactor: all 16 existing tests pass unchanged.
