+++ crates/cli/src/commands/repo/relayout.rs
@@ -1,22 +1,12 @@
-use std::path::PathBuf;
-
 use dm_core::config::Config;
-use dm_core::error::Error as CoreError;
-use dm_core::registry::{RegistryStore, Repo, RepoId};
+use dm_core::relayout::{self, Candidate};
 
 use crate::error::Result;
 use crate::prompt::confirm;
 
-pub struct Candidate {
-    pub id: RepoId,
-    pub name: String,
-    pub from: PathBuf,
-    pub to: PathBuf,
-}
-
 pub fn run(apply: bool, yes: bool) -> Result<()> {
     let config = Config::load()?;
-    let candidates = plan()?;
+    let candidates = relayout::plan()?;
 
     if candidates.is_empty() {
         println!(
@@ -38,7 +28,10 @@ pub fn run(apply: bool, yes: bool) -> Result<()> {
         return Ok(());
     }
 
-    let moved = apply_candidates(candidates)?;
+    let (moved, skipped) = relayout::apply_candidates(candidates)?;
+    for name in &skipped {
+        eprintln!("skipping {name}: target already exists");
+    }
     println!("moved {moved} repo(s)");
     Ok(())
 }
@@ -54,62 +47,9 @@ pub fn print_candidates(candidates: &[Candidate], config: &Config) {
     }
 }
 
-/// Computes which tracked repos (with known host/owner) don't match the
-/// current layout, and where they'd move to.
-pub fn plan() -> Result<Vec<Candidate>> {
-    let config = Config::load()?;
-    let store = RegistryStore::open_default()?;
-    let repos = store.list(None, None)?;
-    Ok(plan_moves(&repos, &config))
-}
-
-/// Moves each candidate on disk and updates its registry entry, skipping
-/// (and warning about) any target that already exists. Returns how many
-/// were actually moved.
-pub fn apply_candidates(candidates: Vec<Candidate>) -> Result<usize> {
-    let store = RegistryStore::open_default()?;
-    let mut moved = 0;
-    for c in candidates {
-        if c.to.exists() {
-            eprintln!("skipping {}: target already exists ({})", c.name, c.to.display());
-            continue;
-        }
-        if let Some(parent) = c.to.parent() {
-            std::fs::create_dir_all(parent).map_err(CoreError::from)?;
-        }
-        std::fs::rename(&c.from, &c.to).map_err(CoreError::from)?;
-        store.update_path(c.id, &c.to)?;
-        moved += 1;
-    }
-    Ok(moved)
-}
-
-fn plan_moves(repos: &[Repo], config: &Config) -> Vec<Candidate> {
-    repos
-        .iter()
-        .filter_map(|repo| {
-            let host = repo.host.as_deref()?;
-            let owner = repo.owner.as_deref()?;
-            let target = config
-                .repo
-                .root
-                .join(config.repo.layout.render(host, owner, &repo.name));
-            if target == repo.path {
-                return None;
-            }
-            Some(Candidate {
-                id: repo.id,
-                name: repo.name.clone(),
-                from: repo.path.clone(),
-                to: target,
-            })
-        })
-        .collect()
-}
-
 /// Whether any tracked repos have host/owner metadata that could drift from
 /// the current layout — used to decide whether to suggest `dm repo
 /// relayout` after `dm config set layout`.
 pub fn has_relayout_candidates() -> Result<bool> {
-    Ok(!plan()?.is_empty())
+    Ok(relayout::has_candidates()?)
 }
