+++ crates/core/src/lib.rs
@@ -5,6 +5,7 @@ pub mod git;
 pub mod layout;
 pub mod paths;
 pub mod registry;
+pub mod relayout;
 pub mod workspace;
 
 pub use error::{Error, Result};
